requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.202
botocore==1.12.203
bz2file==0.98
cachetools==2.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
docutils==0.14
en-core-web-lg==2.1.0
en-core-web-sm==2.1.0
gensim==3.4.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-language==1.3.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
jmespath==0.9.4
joblib==0.13.2
json-tricks==3.13.1
jsonschema==3.0.0a3
kiwisolver==1.1.0
matplotlib==3.1.1
murmurhash==1.0.0
mysql-connector-python==8.0.17
numpy==1.16.4
pandas==0.25.0
plac==0.9.6
preshed==2.0.1
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.2
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
spacy==2.1.4
srsly==0.0.6
thinc==7.0.4
tornado==6.0.3
tqdm==4.32.1
urllib3==1.24.3
wasabi==0.2.2
